restarting sync: support multiple datastores
Removed the limitation to one active datastore. Testing showed that
there was indeed a problem: when processing the Alert which triggers
the restart, the "we are restarting" flag was local and thus only
applied to the first datastore for which such an Alert was
received.

All further Alerts then are processed when the session state was
already reset and thus the other datastores were not reset, causing
undesired slow syncs (anchor not written).

Fixed by remembering the "we are restarting" flag as part of the
session. For good measure the flag is reset again when the restarted
sync enters the "sync" phase.