source link: always allow config options
An app using libsynthesis cannot determine whether the source link
feature was enabled when compiling libsynthesis, but it has to add
the config entries for it to make it work if it was enabled, at
least on Linux ("doxygen" must be chosen there).

This triggers error messages to the console if console printing
is enabled and the source link feature is off.

To avoid that, let's allow the config options and check them
for correctness regardless whether the feature is enabled.

It might also be good to use a sensible default and thus remove
the need to set the options at all, but that is a different issue.