scripting: file URI decoding
Utility method for extracting the file path from an URI.  Needed for
inlining PHOTO data in scripts before sending to a peer.

Uses the existing splitURL() method and enhances it, because the
previous parser was a bit too simplistic. For example, in
file:///tmp/file it skipped all the leading slashes and interpreted
"tmp" as host name, instead of returning an empty host name.

Query and port number are split from the host and doc part directly by
splitURL() now because that seems more consistent.  The in-tree code
using syncURL() concatenates the separate parts to keep its own
semantic unchanged.

URL %xx decoding must be handled by the caller of syncURL(), using the
new urlDecode() utility code. This is not done in syncURL() to
simplify concatenating parts (would have to url encode reserved
characters otherwise).

The new code always returns an empty doc part. There's no longer
a special case for inserting a slash when only the query part is
set, because the caller is already expected to treat the doc name as
starting in the root of the document hierarchy.