SyncCap: X-SYNTHESIS-RESTART -> 390001 (for Funambol)
It turned out the 500 internal server error in the Funambol
server was caused by the non-integer X-SYNTHESIS-RESTART string
in the SyncCap. The server expects integers and throws a parser
exception when sent something else.

The easiest solution is to use a random number that (hopefully)
no-one else uses. The "390" prefix was generated with a random
number generator, "001" is the first extension.