datastore + CTCap: added support for flags
This extends the idea already used for "can restart" to arbitrary
flags: additional sync modes in CTCap can be used to convey additional
information about a datastore to the peer. This only works when using
SyncML 1.2.

The XML "syncmode" element of a <datastore> sets a single flag. It can
occur zero or more times. Each entry is treated as an opaque string by
the engine. For compatibility with SyncML peers which expect sync
modes to be integers, the string should be a valid integer.

All sync modes received from a peer are stored in the TSyncDatastore
when receiving a CTCap. They can be checked with the new
SYNCMODESUPPORTED() script method. This method only works inside
scripts which work with items, like a compare script.