script context: fixed out-of-bounds memory access
TScriptContext::initParse() gets called for all scripts, even if they
are empty. In that case the code read past the end of the string. Now
empty scripts are treated as special case where TScriptContext member
variables are cleared without reading from the script buffer.

This bug was found in valgrind on x86, but not amd64. Didn't cause
problems in practice. Not sure why it didn't show up earlier - perhaps
something in the allocation of std::string changed during a recent
libstdc++ update.