logging: avoid hard-coding log settings when <logsessionstoglobal> is set
Global settings were hard-coded and always used open/close (explicitly
set for server, implied by dbgsubthread_linemix). For SyncEvolution
this is overkill (never shares a log file between session, by setting
a unique log path and not reusing engine instances).

To let SyncEvolution use the desired settings, the overwriting of
values is now conditional on <logsessionstoglobal>: if and only if
set, then session log settings as configured are used, without
changes. If not set the behavior is as before.

This is a slight semantic change. Apps which use <logsessionstoglobal>
now have to set the non-default
<logflushmode>openclose</logflushmode>,
<subthreadmode>mix</subthreadmode> and
<showthreadid>yes</showthreadid> explicitly to get the traditional
behavior.