SyncCap: compatibility enhancement for Nokia phones
It turns out that Nokia phones don't like SyncModes that they don't
recognize. They abort a sync over Bluetooth/OBEX with a generic
"Forbidden" error at the OBEX layer. This only occurs when the sync
app requests the inclusion of non-standard SyncCaps ("can restart",
custom flags).

As a workaround, the extended SyncModes are sent less often now:
- a client always sends them, based on the observation that
  servers handle that okay(ish - see Funambol workaround)
- a server only sends them if the client has already sent
  some extensions itself

This works as long as the Put with the client's DevInf is processed
before the server's DevInf is generated in response to a Get; for
libsynthesis<->libsynthesis syncs that is the case.

A new dummy 390002 SyncMode is always sent, to trigger this behavior
in the server also when the client has no other extensions. Perhaps
this behavior should be made configurable and off by default, to avoid
regressions in combinations which haven't been tested yet.