engine: initial support for caching of remote data
This patch introduces support for true one-way syncing ("caching"):
the local datastore is meant to be an exact copy of the data on the
remote side. The assumption is that no modifications are ever made
locally outside of syncing. This is different from one-way sync modes,
which allows local changes and only temporarily disables sending them
to the remote side.

Another goal of the new mode is to avoid data writes as much as
possible.

This new mode only works on the server side of a sync, where the
engine has enough control over the data flow. It has to be enabled in
an <alertscript> with the new SETCACHEDATA() method, similar to the
way how local read-only mode gets enabled with SETREADONLY(). A
CACHEDATA() query method also gets added.

A SyncML client cannot enable caching mode when alerting the server,
nor can the server alert a client in that mode. The intended usage is
that the caching mode is configured in the layer above the engine and
then gets enabled in the engine after initiating a normal slow or
incremental sync.

Once the mode is enabled, the SyncML server behaves differently as
follows:
- The local sync set is always initialized, even in refresh-from-client
  syncs. That's because the data is intended to be reused during
  refresh-from-client syncs.
- A refresh-from-client is treated like a slow sync: the mapping is
  cleared and remote items are matched.
- After receiving and matching all remote items in a slow sync or
  refresh-from-client sync, any local item that was not matched
  (= its status is still "to be added remotely") gets deleted
  because it no longer exists on the remote side.

The behavior of matching items in caching mode still needs to be
changed to achieve the desired behavior. This is in a separate patch.