engine: ignore unassigned entries at end of array
The engine treated an array with unassigned entries at the end as
larger (and thus different) from an array that ends before those
entries, although reading those entries would also have return
"unassigned".

This caused the engine to treat items as different although there were
semantically identical. Found while testing caching mode with
SyncEvolution, where it is critical to minimize writing to the cache.

The new code checks the larger array to see if all the extra entries
are unassigned, in which case it returns "equal" as result instead of
the previous "larger" or "smaller".