engine: override merge options
In caching mode, the local data is meant to be an exact copy of the
client's data. If the client removes fields in an item and then does a
slow sync, no fields of a matching server item are meant to be
preserved (= merged).

This can be achieved with merge=no in the field list. But the field
list is static and shared between data stores, which might not all
operate in caching mode. This makes updating the config cumbersome.

Therefore this patch adds a flag to the merging functions instead
which overrides the configured behavior such that fields are always
copied from one side into the other. This is activated when in caching
mode and also (optionally) via the MERGEITEMS() function.

The comparison is still needed, to determine whether anything changed
at all during a slow sync and prevent unnecessary database writes.

The loosing side is always the server in caching mode, no need to
check (similar to making the local store read-only).

This all applies to "slow sync" and "conflict" code paths.

In caching mode, a <mergescript> is currently ignored in favor of
doing the intended copying directly. Needs to be fixed.