SyncML TK: don't read past end of buffer
The code which dumps the buffer after a parsing error had no way
of obtaining the buffer size and thus read past the end of the buffer,
as seen in SyncEvolution nightly testing under valgrind.

Must provide the buffer length together with a pointer. The extra
information is optional, so only the places which needed to know
the length were updated.