engine: fix thread support for Linux
This allows enabling multithreading on Linux, without enabling it yet.
Linux uses pthread.h for threads, like Android, but in contrast to Android,
we have to define uIntArch ourselves.

On 64 bit Linux, that type has to be 64 bit wide and won't fit into
the uInt32 that was used in some places. Some type casts were added to
print statements, to avoid mismatches between the format string and
the actual uIntArch type.

A NULL was necessary because at one point, installOutput() was called
when there was no fDbgOutP, thus causing a crash on the
fDbgOutP->clone() call.