Linux: allow overriding blocking pthread code
Blocking the main thread inside pthread_cond_(timed)wait() for long
periods of time is bad when that thread also needs to process events.

For example, in SyncEvolution the main thread must continue to check for
events on the glib event loop, otherwise the background thread can get stuck
in EDS and/or glib while waiting for that event processing.

As a stop-gap measure, export a function pointer that the user of the
Linux libsynthesis can override. The replacement must check for thread
completion and be able to use the mutex+condition variable for this,
so pass all of those as parameters.

A complete solution would be to define a plugable threading system
where the user of the library can replace the entire thread start/stop
with its own code.