debuglogger: fix uninitialized memory read when using separate logs
valgrind complained about an uninitialized memory read of fBlockNo
after enabling multithreading with separate sublogs.

Initializing in the constructor fixes the problem. Hopefully this
is the right fix; I'm not entirely sure under which circumstances
the member was meant to be set.

Also set fGZonesP, which was another member not set by the constructor.
There was no valgrind report about that, though.