asynchronous command execution
It can be useful to trigger add/update/delete requests and check
for completion later. For example, combining multiple updates
in one operation is more efficient in the Evolution Data Server
(used by SyncEvolution). Another advantage is the execution of
these changes in the background while already exchanging the
next message with the peer.

This patch implements asynchronous add/update/delete operations for
client and server. The functionality can be requested by data stores
by returning the new LOCERR_AGAIN status code.

The <strictexecordering> flag must be set to 'false' before multiple
pending operations are allowed. It is still 'true' by default, in
which case once the first operation returns LOCERR_AGAIN, all further
operations are queued immediately without submitting them to the store
first.

The flag used to allow processing further commands when the first Sync
command was queued, primarily (exclusively?) because of a slow data
store 'start data read' operation. Testing showed that allowing that
breaks syncing with the Synthesis engine, so instead of allowing
something of dubious value the code now always serializes commands
when anything more than complex than add/delete/replace/copy is
involved.

In that sense the commit reduces the semantic of <strictexecordering>
slightly from "allow broken SyncML behavior" to "enable asynchronous
command execution". Together with the next commit, that hopefully can
be done without violating the SyncML spec.

Regarding the implementation, When getting the LOCERR_AGAIN status
back, all levels involved in the processing of the change dump their
local state into some kind of ausiliary data structure and attach that
to the item. They check for that auxiliary data the next time they get
called and if found, know that they need to resume operation where
they left of before.

For small functions, this is done with changes the if/else code
branches. For large functions, the changes would have been too
intrusive and hard to verify for correctness, because the control flow
often depends on variables that may change as part of the
execution. An approach based on goto was used in those cases (kind of
like a reverse long jump).

TSyncItem stores the auxiliary data in a hash indexed with hard-coded
keys for each layer. Arbitrary string keys were considered, but
discarded because of the (unnecessary) performance overhead. The
SyncML toolkit's SmlItem has a new pointer that holds the additional
state. The toolkit itself only sets it and knows how to free it via
a callback.

There is no explicit "flush pending changes" or "complete changes"
operation on data stores, because that would require extending the
plugin API. Instead the data stores are expected to finish the item
change(s) when called a second time. Normally the second call happens
at the beginning of the next message processing, thus achieving the
goal of overlapping processing with network IO.

The exception is the last message in a session. Testing showed that
delaying item changes in the SyncML client's final message confused
Synthesis client and/or server so that they no longer agree on the end
of session. Now the pending commands are issued a second time at the
end of processing that final message, which solved that issue.

There is one caveat that needs to be addressed in the next commit:
statuses must be issued in the order of their commands. At the moment,
the engine in its default "strict exec ordering" mode will stop
issuing more commands after the initial one was queued. A different
way of ensuring correct order will be needed.