ensure ordering of status replies
This commit ensures that Status commands are sent in the order of
the corresponding command when out-of-order execution of commands
is enabled.

This is achieved by checking whether a Status command can really be
issued. If there are already commands queued, then any Status command
issued while processing another command gets attached to that
command. Then the command itself is queued although it has
finished. From then on the command merely acts as a container for the
status commands.

The next time delayed commands are processed, such commands are not
executed again. Instead their statuses are issued.

Add/Replace commands which contain incomplete items are special: they
must trigger a Status 213 response in the same message, which cannot
be done without processing the previously queued commands first.

We could try to queue the incomplete command without processing it at
all, but that would change the response in a way that might be
unexpected by the peer (although legal). Better play it safe and
finish all queued commands, then process the command.

Out-of-order execution still needs to be enabled
explicitly via <strictexecordering>off</strictexecordering> in
a remote rule. This is a bit misleading, because the behavior
should not be visible to the peer anymore, and thus not depend
on who we are talking to.