engine: fix MERGEFIELDS()
Commit de7ff9 introduced a new mode of operation for merging, where
one side gets updated to be identical to the other side, regardless of
what the field list says.

That commit broke MERGEFIELDS() is several ways:
- The extra parameter was not declared in TBuiltInFuncDef
  and therefore all attempts to call MERGEFIELDS() from
  a script caused a parsing error (something like "closing
  bracket expected").
- The parameter was meant to be optional, but not passing it
  caused a crash because aFuncContextP->getLocalVar(0) then
  is NULL.

This commit fixes these issues by making the new parameter optional
and checking whether it exists (probably not necessary) and was set.