engine: <mergescript> + caching mode
When in caching mode, the purpose of merging is to make the server
item identical to the client item and detecting when the server item
was changed. Doing this in combination with a merge script was not
possible.

Now the extra mode parameter is passed through standardMergeWith()
into the merge script, where it can be retrieved via the new
MERGEMODE() method.

0 is the traditional mode which updates both items. 1 updates the
loosing item (this is the mode used for caching) and 2 updates the
winning item (was added for the sake of completeness).