linux: fix multithreading under valgrind
TThreadObject::waitfor() gets called twice: once when the engine waits for
completion of the thread, once when resetting the datastore.  The second wait
should be a NOP. However, although skipped the condwait (because
fTerminated==true) it still ended up calling pthread_join() again (because
fPosixThread!=0).

This is problematic because a thread can only be joined once. It happened to
work most of the time, except in some tests running under valgrind, where the
second pthread_join() hangs. It is uncertain why it didn't happen all the time
(Number of threads? Runtime of program?), but the fix is obvious - prevent
entering pthread_join() more than once by resetting fPosixThread after the
first call.