Plugin: allow overriding support for suspend/resume
Previously, the support for suspend/resume in a plugin DB was tied to
the version of its API. This is a bit inflexible, because the API that
the DB provides is orthogonal to the behavior of its
implementation.

For example, SyncEvolution always provides the same API, but sometimes
decides to not store any data across sync sessions (caching mode which
will always start anew as slow sync). In that mode, disabling
suspend/resume support speeds up a sync involving 10000 items from 38s
to 23s.

To allow disabling suspend/resume, the new plugin capability
CA_ResumeSupported can be set to "no" in the capability string.

A different implementation with a datastore config option would also
have been possible, but this approach was the least intrusive and
easiest.