Linux: speed up getSystemNowAs(TCTX_SYSTEM)
getSystemNowAs(TCTX_SYSTEM) is called often as part of logging. The
showed up a one of the hotspots during syncing. The TzOffsetSeconds()
showed up as one of the hotspots.

Using localtime_r() and the glibc tm_gmtoff extensions improved
runtime of a sync with a fairly low amount of logging from 25s to 22s.