engine: fix LOCERR_AGAIN handling with strict exec ordering
Using SyncEvolution as SyncML client was basically broken: when receiving
a message with a few Adds and Final, the first Add returned LOCERR_AGAIN.
This caused the engine to queue the command and all following ones, without
invoking the storage for those additional queued Adds. So batching of writes
was not working. Worse, the map phase was entered and thus the write phase
ended despite the pending Adds.

As a quick fix, force the storage to execute the operation immediately.

A better solution would be to allow item operations to queue up and then
flush them at the end of the message processing. This would at least allow
batching inside the same message.