binfilebase: avoid virtual methods in destructor
cppcheck complains about the virtual method call to platformFileIsOpen that
happens indirectly inside the destructor if the instance has not been
deconstructed. Because all derived classes must deconstruct the base class
first (mentioned in the binfilebase comments), the base class destructor
should never have to do anything. If it did, the program would crash.

Therefore it is better to not even attempt to do anything in the destructor.