syncsession: track pointer
clang's scan-build reported a "use after free" error. Probably
a false positive, caused by clang not knowing that a function
call will always return the same value. Nevertheless it is safer
to really track whether the pointer is still non-null, in particular
because there are exception handlers which might do a double free.