binfileclient: support sending NumberOfChanges
In a binfileclient, the NumberOfChanges value became available to late
to get included in the TSyncCommand when constructing that
command. Even when it got calculated later in
TBinfileImplDS::changeLogPreflight() it never got sent to the server.

Making the value available earlier is not possible (changeLogPreflight
depends on information calculated after generating the sync command),
but it is possible to re-check the NumberOfChanges before continuing
the sync in the next message. The result is a sync where the first
several items are sent without NumberOfChanges, then the next message
has NumberOfChanges.