debuglogger: close log file on exec
When using fork+exec on Linux, open files get inherited by the child
process unless explicitly marked as FD_CLOEXEC. We should set that
flag at least for files which remain open (like the log file when
dbgflush_openclose is not active). A truly thread-safe approach would
have to use open(O_CLOEXEC), but setting the flag later is sufficient
for us.

This change is limited to Linux to avoid interfering with other
platforms.