profile: <parameter sharedfield>
Traditionally, the value of a property and its parameter values get
stored in fields reserved for that property such that there is no
overlap with other properties. In such a scenario it is enough to
check for unused fields used for the property value, because the
corresponding parameter fields are guaranteed to be available once a
repetition is found.

However, once the same field is used for different properties the
check and creation becomes more complicated: when parsing properties,
the shared field must also be checked. It is not enough for it to be
empty, it also must not be in use at all, because otherwise a value might
get added to the repetition used already by some other property.

This check only makes sense (and only works) for array fields. It is
disabled by default and needs to be enabled explicitly with the new
parameter attribute "sharedfield".

Storing values must ensure that the shared array field gets resized if
the parameter was not set, otherwise the check would fail to detect
used repetitions.

The effect of using a shared field is that arrays contain a lot of
unassigned values because index overlaps between otherwise unrelated
properties get avoided.

This feature was needed for converting between a profile with custom
labels in a X-ABLabel property attached to various contact properties
(ADR, TEL, X-ABDate, etc.) via group tags and a profile using a
property parameter instead. The label array field is the shared field
in this case.