mimedir parser: avoid reusing group tag
When two properties (TEL and X-ABLabel in this example) share the same group
field, then storing the property where a group tag was used for the first time
must not reuse unassigned group field entries.

There was some logic for that already (see "someGroups"), but it failed to
handle this case. This approach here uses brute-force instead of trying to me
smart: the reuse check in the n<maxrep for loop will fail repeatedly at the
beginning until reaching the end of the group field array.