ODBC: avoid clang warning
clang 3.4 points out that the ! in !aQuoteMode!=qm_none in
quoteStringAppend() only gets applied to aQuoteMode. Explicit
brackets silence the compiler warning without changing the
expression.

The compiler indeed points out some unusual code. The if branch is
taken for aQuoteMode == qm_none (probably what was intended) because
!qm_none != qm_none with qm_none == 0 and not taken for all other
values where !aQuoteMode is false and thus equal to qm_none.

It probably would be simpler to write "if (aQuoteMode == qm_none)" but
I don't want to change the meaning of the code.