binfile: /dev/null datadir path
When the client already knows that it is never going to sync again or
the next sync will have to be a slow sync anyway, then the binfile
client files are not needed. Suppressing writing of them is useful on
devices with flash memory to avoid flash wearout. The specific use
case is PBAP caching in a car's head unit.

When setting "/dev/null" as datadir, the platform layer will avoid
creating the files and instead keep their content in memory. This is
required (and sufficient) for syncing once, which will close and later
read from some files at least once. Therefore the data cannot be
thrown away completely.