SMLTK: call memory functions through pointers
This additional indirection allows users of the library to replace the
default implementation from libc.

The feature is optional. In libsynthesis it can be enabled in
target_options.h. Users of libsynthesis do not have access to
libmem.h, so they need to provide their own declaration of the
function pointers.