binfile: avoid unnecessary write and truncate
The goal is to leave binfile completely untouched, including its "last
modified" time stamp, if nothing changes.

There are cases where the upper layers will call
TBinFileBase::updateRecord() and write the same data that already
is in the file. Instead of catching those case-by-case, let's check
at the lowest level.

On btrfs (and possibly other file systems), ftruncate() bumps the
"last modified" time stamp of a file even if the requested size
matches the existing one. Avoid that by skipping the ftruncate() when
the file already has the desired size.