merge scripts: new RELAXEDCOMPARE(), enhanced MERGEFIELDS()
Caching mode in SyncEvolution only worked if the incoming item matched
the stored one exactly. Otherwise changing the order of items
(e.g. TEL followed by EMAIL instead of EMAIL followed by TEL) caused
arrays to be different (because of the shared LABEL array), and that
difference in the arrays led to rewriting items even if nothing
changed.

Because of the decode/encode step before storing a new item, caching
only worked reliably when the incoming item was also decoded and
encoded once, which is unnecessary overhead and only worked when using
a storage with the same encoding as the local sync.

A better solution is to use a more intelligent merge script:
- check whether arrays are identical except for ordering (done
  with the new RELAXEDCOMPARE())
- ignore arrays which were found to be similar enough (by
  passing a new parameter listing these fields to MERGEFIELDS().

In essence, MERGEFIELDS() then is only asked to check the remaining
fields or merge arrays if relevant differences were found.

At the moment, RELAXEDCOMPARE() only checks that entries in the arrays
match. It does not check that their order is the same. This could
(should!) be added and would also make the comparison faster when
there are differences (because it only needs to compare the next
non-empty entries).