Linux time zones: modernize libical support
libical v2 introduces a new API for generating system time zones
in a more interoperable format, with rules for summer and winter time (same
as in libical before v1) instead of just the transition dates (as in
libical v1). We should use that whenever possible.

We can drop support for libical embedded in libecal, that's really ancient.

Binary compatibility is now enabled with a dedicated configure parameter
(cleaner that piggy-backing on the parameter used by SyncEvolution). To make
this work for the libical v1/v2 ABI break (icalarray changed its types),
we must wrap some calls.