sysync_utils: fix clang warnings
Clang 4.8 started warning about unused static variables. ASCIIfyTable is
indeed only used under an ifdef. That can be resolved by moving the table next
to the code.

But Ansi_80_to_9F_to_UCS4 is used, and removing it leads to an undefined
symbol error. This seems to be a false positive. The only way to suppress
it is to add an external declaration.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>